pyproject.toml
[project]
name = "vkusnie-cookies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler[asyncio]>=3.11.2",
    "celery[redis]>=5.6.2",
    "fastapi>=0.128.8",
    "isort>=7.0.0",
    "playwright>=1.58.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-rucaptcha>=6.5.0",
    "redis>=6.4.0",
    "ruff>=0.15.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
